cambios backend y sistema de la app

+++ database/migrations/2026_07_24_173433_create_role_permission_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('role_permission')) {
+            return;
+        }
         Schema::create('role_permission', function (Blueprint $table) {
             $table->id();
             $table->foreignId('role_id')->constrained('roles')->onDelete('cascade');
